Added showCurrentDataWhileFetching option
The map markers were blinking every time the map was moved. This was because every time the map moves it changes bounds in the URL search params. Changes to search params trigger our React Query hook to fetch new data. This in turn causes the ListingMarker components to re-render. There was no amount of memoization that I could do which would prevent them from re-rendering. There was also no way I could find to keep the existing AdvancedMarker instead of creating a new one on each render. The only solution that has worked is to use React Query options placeholderData attribute with the keepPreviousData function. This simply adds the data from the last fetch as a "placeholder" while the new data is fetched.