Switched to GeoJSON for map boundary
The Google Maps API has a data layer API that can be used to add objects to the map that are in GeoJSON format. Decided to switch from using the Polygon class for the map boundary to using this data layer format instead. It seemed like a natural fit since the