Adjusting this to fix the index signature error again
I want ParamDefaults to show up as a constant with only one possible value for each property. This way hovering over its properties will reveal what value it has when I use it inside of a component. Setting its type to SearchState is technically correct, and solves my indexing issue below, but at the expense of not revealing the specific values it has in all the places where I use it. Fixing the indexing issue doesn't really give us much of anything in terms of useful type safety. It's really just a way to fix an annoying Typescript error. I would rather get useful feedback than jump through hoops to satisfy a compiler so we're just going to cast this key when indexing it and call it a day.